新增火焰球音效，修复错位bug

+++ mario_fire.h
@@ -16,5 +16,5 @@ public:
 	std::pair<double, double> getctpos();
 	bool update();
 	void render(double x, double y);
-	Mario_fire(int x, int y, int direction);
+	Mario_fire(double x, double y, int direction);
 };
